+++ app/build.gradle
@@ -45,10 +45,7 @@ ext {
     constraint_layout_version = '1.1.0-beta5'
     ktx_version = '0.1'
 
-    firebase_version = '11.8.0'
-
     rxkotlin_version = '2.2.0'
-    rxlifecycle_version = ''
 
     kodein_version = '4.1.0'
 
@@ -62,9 +59,6 @@ dependencies {
 
     // kotlin
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
-
-    // firebase
-    implementation "com.google.firebase:firebase-auth:$firebase_version"
 
     // core
     implementation "com.android.support:appcompat-v7:$support_lib_version"
