--- app/gradle.properties
@@ -1,2 +1,0 @@
-CLIENT_ID="d1b23a463ecd868bb70b"
-CLIENT_SECRET="2998edf778bbe068150b05f6b9f71a484ab2488e"
