+++ app/src/main/java/coding_trip/io/android/BaseApplication.kt
@@ -1,16 +1,6 @@
 package coding_trip.io.android
 
 import android.app.Application
-import com.google.firebase.FirebaseApp
 
 class BaseApplication : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-        setupFirebase()
-    }
-
-    private fun setupFirebase() {
-        FirebaseApp.initializeApp(this)
-    }
 }
