--- app/src/main/java/coding_trip/io/android/datasource/UserDataSource.kt
@@ -1,5 +1,0 @@
-package coding_trip.io.android.datasource
-
-class UserDataSource {
-
-}
