--- app/src/main/java/coding_trip/io/android/repository/FirebaseRepository.kt
@@ -1,16 +1,0 @@
-package coding_trip.io.android.repository
-
-import coding_trip.io.android.datasource.FirebaseDataSource
-
-class FirebaseRepository (
-    private val dataSource: FirebaseDataSource
-) {
-
-    fun setupUser(accessToken: String) {
-
-    }
-
-    fun updateProfileImageUrl() {
-
-    }
-}
