--- app/src/main/java/coding_trip/io/android/repository/UserRepository.kt
@@ -1,9 +1,0 @@
-package coding_trip.io.android.repository
-
-import coding_trip.io.android.datasource.UserDataSource
-
-class UserRepository(
-    private val dataSource: UserDataSource
-) {
-
-}
