+++ app/src/main/java/coding_trip/io/android/ui/start/StartActivity.kt
@@ -8,11 +8,8 @@ import android.support.v7.app.AppCompatActivity
 import android.util.Log
 import coding_trip.io.android.BuildConfig
 import coding_trip.io.android.R
-import coding_trip.io.android.datasource.AuthDataSource
 import coding_trip.io.android.repository.AuthRepository
 import coding_trip.io.android.ui.home.HomeActivity
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.GithubAuthProvider
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
 import kotlinx.android.synthetic.main.activity_start.login
@@ -41,49 +38,30 @@ class StartActivity : AppCompatActivity() {
     }
 
     private fun openCustomTabs() {
-        val url = getString(R.string.str_oath_url, BuildConfig.CLIENT_ID, SCOPE)
+        val url = "https://github.com/login/oauth/authorize" +
+            "?client_id=${BuildConfig.CLIENT_ID}&scope=$SCOPE"
         val customTabsIntent = CustomTabsIntent.Builder().build()
         customTabsIntent.launchUrl(this, Uri.parse(url))
     }
 
     private fun getAccessToken(code: String) {
         // this part may be replaced after implementing kodein
-        val authDataSource = AuthDataSource(this)
-        val authRepository = AuthRepository(authDataSource)
+        val authRepository = AuthRepository(this)
 
-        authRepository.fetch(code)
+        authRepository.fetchAccessToken(code)
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe({
-                authRepository.saveAuthToken(it)
+                authRepository.saveAccessToken(it)
+                Log.d("StartActivity", "Token: $it")
 
                 val intent = Intent(this, HomeActivity::class.java)
                 startActivity(intent)
-//                createFirebaseUser(it)
             }, {
                 TODO("This log implementation will be replaced by Timber")
                 Log.e("StartActivity", "Failed to load access token")
             })
     }
-
-    /*
-    private fun createFirebaseUser(token: String) {
-        val credential = GithubAuthProvider.getCredential(token)
-        val auth = FirebaseAuth.getInstance()
-        auth.signInWithCredential(credential)
-            .addOnCompleteListener {
-                if (it.isSuccessful.not()) {
-                    Log.e("StartActivity", "Failed to authorized")
-                } else {
-                    val intent = Intent(this, HomeActivity::class.java)
-                    startActivity(intent)
-                }
-            }
-            .addOnFailureListener {
-                Log.e("StartActivity", "Failed to authorized")
-            }
-    }
-    */
 
     companion object {
         const val SCOPE = "user"
